+++ sdk/datafactory/azure-mgmt-datafactory/azure/mgmt/datafactory/_data_factory_management_client.py
@@ -7,17 +7,19 @@
 # --------------------------------------------------------------------------
 
 from copy import deepcopy
-from typing import Any, TYPE_CHECKING
+from typing import Any, Optional, TYPE_CHECKING, cast
 from typing_extensions import Self
 
 from azure.core.pipeline import policies
 from azure.core.rest import HttpRequest, HttpResponse
+from azure.core.settings import settings
 from azure.mgmt.core import ARMPipelineClient
 from azure.mgmt.core.policies import ARMAutoResourceProviderRegistrationPolicy
+from azure.mgmt.core.tools import get_arm_endpoints
 
 from . import models as _models
 from ._configuration import DataFactoryManagementClientConfiguration
-from ._serialization import Deserializer, Serializer
+from ._utils.serialization import Deserializer, Serializer
 from .operations import (
     ActivityRunsOperations,
     ChangeDataCaptureOperations,
@@ -30,6 +32,7 @@ from .operations import (
     GlobalParametersOperations,
     IntegrationRuntimeNodesOperations,
     IntegrationRuntimeObjectMetadataOperations,
+    IntegrationRuntimeOperations,
     IntegrationRuntimesOperations,
     LinkedServicesOperations,
     ManagedPrivateEndpointsOperations,
@@ -45,6 +48,7 @@ from .operations import (
 )
 
 if TYPE_CHECKING:
+    from azure.core import AzureClouds
     from azure.core.credentials import TokenCredential
 
 
@@ -67,6 +71,8 @@ class DataFactoryManagementClient:  # pylint: disable=too-many-instance-attribut
     :ivar integration_runtime_nodes: IntegrationRuntimeNodesOperations operations
     :vartype integration_runtime_nodes:
      azure.mgmt.datafactory.operations.IntegrationRuntimeNodesOperations
+    :ivar integration_runtime: IntegrationRuntimeOperations operations
+    :vartype integration_runtime: azure.mgmt.datafactory.operations.IntegrationRuntimeOperations
     :ivar linked_services: LinkedServicesOperations operations
     :vartype linked_services: azure.mgmt.datafactory.operations.LinkedServicesOperations
     :ivar datasets: DatasetsOperations operations
@@ -112,8 +118,11 @@ class DataFactoryManagementClient:  # pylint: disable=too-many-instance-attribut
     :type credential: ~azure.core.credentials.TokenCredential
     :param subscription_id: The subscription identifier. Required.
     :type subscription_id: str
-    :param base_url: Service URL. Default value is "https://management.azure.com".
+    :param base_url: Service URL. Default value is None.
     :type base_url: str
+    :keyword cloud_setting: The cloud setting for which to get the ARM endpoint. Default value is
+     None.
+    :paramtype cloud_setting: ~azure.core.AzureClouds
     :keyword api_version: Api Version. Default value is "2018-06-01". Note that overriding this
      default value may result in unsupported behavior.
     :paramtype api_version: str
@@ -125,12 +134,24 @@ class DataFactoryManagementClient:  # pylint: disable=too-many-instance-attribut
         self,
         credential: "TokenCredential",
         subscription_id: str,
-        base_url: str = "https://management.azure.com",
+        base_url: Optional[str] = None,
+        *,
+        cloud_setting: Optional["AzureClouds"] = None,
         **kwargs: Any
     ) -> None:
+        _cloud = cloud_setting or settings.current.azure_cloud  # type: ignore
+        _endpoints = get_arm_endpoints(_cloud)
+        if not base_url:
+            base_url = _endpoints["resource_manager"]
+        credential_scopes = kwargs.pop("credential_scopes", _endpoints["credential_scopes"])
         self._config = DataFactoryManagementClientConfiguration(
-            credential=credential, subscription_id=subscription_id, **kwargs
+            credential=credential,
+            subscription_id=subscription_id,
+            cloud_setting=cloud_setting,
+            credential_scopes=credential_scopes,
+            **kwargs
         )
+
         _policies = kwargs.pop("policies", None)
         if _policies is None:
             _policies = [
@@ -149,7 +170,7 @@ class DataFactoryManagementClient:  # pylint: disable=too-many-instance-attribut
                 policies.SensitiveHeaderCleanupPolicy(**kwargs) if self._config.redirect_policy else None,
                 self._config.http_logging_policy,
             ]
-        self._client: ARMPipelineClient = ARMPipelineClient(base_url=base_url, policies=_policies, **kwargs)
+        self._client: ARMPipelineClient = ARMPipelineClient(base_url=cast(str, base_url), policies=_policies, **kwargs)
 
         client_models = {k: v for k, v in _models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
@@ -167,6 +188,9 @@ class DataFactoryManagementClient:  # pylint: disable=too-many-instance-attribut
             self._client, self._config, self._serialize, self._deserialize
         )
         self.integration_runtime_nodes = IntegrationRuntimeNodesOperations(
+            self._client, self._config, self._serialize, self._deserialize
+        )
+        self.integration_runtime = IntegrationRuntimeOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
         self.linked_services = LinkedServicesOperations(self._client, self._config, self._serialize, self._deserialize)